[z3tracer] rename events -> syntax
