+++ z3tracer/src/error.rs
@@ -1,7 +1,7 @@
 // Copyright (c) Facebook, Inc. and its affiliates
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
-use crate::events::Ident;
+use crate::syntax::Ident;
 
 #[derive(Debug)]
 pub enum Error {
